update: error handling

+++ backend/todo-backend/src/service/todo.ts
@@ -1,12 +1,15 @@
 import * as todoRepo from '../repos/todo';
 import { Todo } from '../model/todo';
+import { NotFoundError } from '../error/notFoundError';
 
 export const getAllTodos = () => {
   return todoRepo.getAllTodos();
 };
 
 export const getUserTodos = (id: number) => {
-  return todoRepo.getUserTodos(id);
+  const todos = todoRepo.getUserTodos(id);
+  if (!todos) throw new NotFoundError('Todos not found');
+  return todos;
 };
 
 export const addTodo = (todo: Todo) => {
@@ -14,13 +17,19 @@ export const addTodo = (todo: Todo) => {
 };
 
 export const getTodoById = (id: number, userid: number) => {
-  return todoRepo.getTodoById(id, userid);
+  const todos = todoRepo.getTodoById(id, userid);
+  if (!todos) throw new NotFoundError('Todo not found');
+  return todos;
 };
 
 export const updateTodo = (id: number, todo: Todo) => {
+  const foundTodo = todoRepo.getTodoById(id, todo.userid);
+  if (!foundTodo) throw new NotFoundError('Todo not found');
   return todoRepo.updateTodo(id, todo);
 };
 
-export const deleteTodo = (id: number) => {
+export const deleteTodo = (id: number, userid: number) => {
+  const todo = todoRepo.getTodoById(id, userid);
+  if (!todo) throw new NotFoundError('Todo not found');
   return todoRepo.deleteTodo(id);
 };
